# internal/api/openapi.yaml
basePath: /api
definitions:
...
  api.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.Principal'
    type: object
  api.bindingEdge:
    properties:
      environment:
        type: string
      from:
        type: string
      to:
        type: string
    type: object
...
  api.deployRequest:
    properties:
      environment:
        type: string
      image:
        type: string
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.setupRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  api.statusResponse:
    properties:
      setupRequired:
        type: boolean
    type: object
...
  auth.Principal:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/auth.Role'
    type: object
  auth.Role:
    enum:
    - admin
    - member
    - viewer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
    - RoleViewer
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted Railway-style deploy platform for Kubernetes. Manages projects,
    apps, deployments, env vars, secrets, domains, and git providers.
  title: Mortise API
  version: 0.1.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Log in
      tags:
      - auth
  /auth/setup:
    post:
      consumes:
      - application/json
      description: Creates the first admin user. Returns 409 if any user already exists.
      parameters:
      - description: Admin credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.setupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: First-time setup
      tags:
      - auth
  /auth/status:
    get:
      description: Reports whether first-user setup is required (no users exist yet)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Check setup status
      tags:
      - auth
...
  /projects/{project}/apps/{app}/deploy:
    post:
      consumes:
      - application/json
      description: Trigger a deploy by patching the App's source image. Accepts either
        a JWT (user session) or a deploy token (CI) for authentication.
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      - description: Deploy details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.deployRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Deploy an app
      tags:
      - deploy
...
  /projects/{project}/bindings:
    get:
      description: Returns all binding edges (from-app to to-app) for the specified
        environment
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: Environment name
        in: query
        name: environment
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.bindingEdge'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: List bindings for a project environment
      tags:
      - bindings
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token from POST /api/auth/login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"